requirements.txt
crispy-bootstrap5==2024.10
Django==5.1.4
django-colorfield==0.11.0
django-crispy-forms==2.3
graphviz==0.20.3
